backend/pyproject.toml
[tool.poetry]
name = "fydp"
version = "0.1.0"
description = ""
authors = ["josh folkins"]
readme = "README.md"

[tool.poetry.dependencies]
python = ">=3.11,<3.13"
python-docx = "^1.1.2"
transformers = "^4.42.3"
torch = "^2.3.1"
tqdm = "^4.66.4"
chromadb = "^0.5.4"
sentence-transformers = "^3.0.1"
flagembedding = "^1.2.10"
wheel = "^0.43.0"
einops = "^0.8.0"
semantic-router = "^0.0.51"
python-dotenv = "^1.0.1"
pysqlite3-binary = "^0.5.3"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
